backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.4.0
gunicorn==22.0.0
whitenoise==6.7.0
drf-nested-routers==0.94.1
